requirements.txt
fastapi
uvicorn
pydantic
motor
email-validator
python-decouple
starlette
datetime
python-multipart
PyJWT
passlib